style, move `config.ts`
